requirements2.txt
python-dotenv
langchain_groq
langchain_core
langchain
mysql-connector-python
sqlalchemy
pymysql
langchain-community